chore: fix folia build pipeline
