+++ orebfuscator-compatibility/orebfuscator-compatibility-paper/src/main/java/net/imprex/orebfuscator/compatibility/paper/PaperCompatibilityLayer.java
@@ -22,11 +22,6 @@ public class PaperCompatibilityLayer implements CompatibilityLayer {
 	}
 
 	@Override
-	public boolean isMojangMapped() {
-		return true;
-	}
-
-	@Override
 	public CompatibilityScheduler getScheduler() {
 		return this.scheduler;
 	}
