+++ orebfuscator-plugin/pom.xml
@@ -13,7 +13,6 @@
 	<packaging>jar</packaging>
 	
 	<build>
-		<finalName>${name}-${version}</finalName>
 		<resources>
 			<resource>
 				<directory>src/main/resources</directory>
@@ -105,6 +104,24 @@
 		<dependency>
 			<groupId>net.imprex</groupId>
 			<artifactId>orebfuscator-compatibility-api</artifactId>
+			<version>${revision}</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-compatibility-bukkit</artifactId>
+			<version>${revision}</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-compatibility-folia</artifactId>
+			<version>${revision}</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-compatibility-paper</artifactId>
 			<version>${revision}</version>
 			<scope>compile</scope>
 		</dependency>
@@ -208,8 +225,22 @@
 		</dependency>
 		<dependency>
 			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-nms-v1_19_R1</artifactId>
+			<version>${revision}</version>
+			<classifier>mojang-mapped</classifier>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
 			<artifactId>orebfuscator-nms-v1_19_R2</artifactId>
 			<version>${revision}</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-nms-v1_19_R2</artifactId>
+			<version>${revision}</version>
+			<classifier>mojang-mapped</classifier>
 			<scope>compile</scope>
 		</dependency>
 		<dependency>
@@ -220,8 +251,22 @@
 		</dependency>
 		<dependency>
 			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-nms-v1_19_R3</artifactId>
+			<version>${revision}</version>
+			<classifier>mojang-mapped</classifier>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
 			<artifactId>orebfuscator-nms-v1_20_R1</artifactId>
 			<version>${revision}</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>net.imprex</groupId>
+			<artifactId>orebfuscator-nms-v1_20_R1</artifactId>
+			<version>${revision}</version>
+			<classifier>mojang-mapped</classifier>
 			<scope>compile</scope>
 		</dependency>
 	</dependencies>
